refactor: move testDto files to new directory
