+++ parser/src/main/java/com/apighost/parser/reader/ScenarioReader.java
@@ -1,6 +1,6 @@
 package com.apighost.parser.reader;
 
-import com.apighost.model.scenario.Scenario;
+import com.apighost.model.scenario.testfile.Scenario;
 
 public interface ScenarioReader {
     public Scenario readScenario ();
